implement size_t for utility/algo/helper classes

+++ utility.h
@@ -51,7 +51,7 @@ namespace internal
 	template<typename T> RDE_FORCEINLINE
 	void move_n(const T* from, size_t n, T* result, int_to_type<false>)
 	{
-		for (int i = int(n) - 1; i >= 0; --i)
+		for (size_t i = n - 1; i >= 0; --i)
 			result[i] = from[i];
 	}
 	template<typename T> RDE_FORCEINLINE
@@ -161,7 +161,7 @@ namespace internal
 			new (to + i) T();
 	}
 	template<typename T> inline
-	void construct_n(T*, int, int_to_type<true>)
+	void construct_n(T*, size_t, int_to_type<true>)
 	{
 		// trivial ctor, nothing to do.
 	}
